Fixed circular import error in vault addresses subscription
