+++ vouchers/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class VoucherConfig(AppConfig):
     name = 'vouchers'
+
+    def ready(self):
+        import vouchers.signals
